Add files via upload
What it does: It makes sound files easier to use and sort. It changes the sound files from one type to another type that is more common. It also puts the sound files into different folders based on their names.
How it works: It uses a special language called Python to read the sound files, make them louder or quieter, and change their type. Then it makes new folders and moves the sound files into them. The folders have names that match the first three letters of the sound files.
Why it is useful: It helps people who work with sound files to make them better and more organized. It saves time and effort. It can be changed to fit different needs.